+++ ChangeLog.md
@@ -3,7 +3,7 @@ CHANGE LOG
 -------------------
 
 --------
-201709XX  Not yet released
+20170913
 --------
 
 1.7.4
@@ -17,8 +17,8 @@ Some refactoring of Patron and Valet
 Fixed windows bug in serialing
 change Share so .update or .create when passed in another Share works
 Fixed issue with using transmit with Patron if waiting on request/response.
-Refactored to have new .request method that uses existing parameters to generate
-new request and append  to .requests this then honors .waiting flag
+Refactored to add Patron.request method that uses existing parameters to generate
+   a new request and append  to .requests. This honors .waiting flag
 
 --------
 20170830